pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = ["newxxson <hyukjun1111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.4"
sqlalchemy = "^2.0.31"
ruff = "^0.5.5"
mypy = "^1.11.1"
pytest = "^8.3.2"
pydantic = "^2.8.2"
python-json-logger = "^2.0.7"
injector = "^0.22.0"
alembic = "^1.13.2"
pre-commit = "^3.8.0"
confluent-kafka = "^2.5.0"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pytest-asyncio = "^0.23.8"
pytest-cov = "^5.0.0"
msgspec = "^0.18.6"
geoalchemy2 = "^0.15.2"
aiohttp = "^3.10.1"
hypothesis = "^6.110.1"
python-dateutil = "^2.9.0.post0"
shapely = "^2.0.5"
gunicorn = "^23.0.0"
pyjwt = "^2.9.0"
crypto = "^1.4.1"
python-multipart = "^0.0.17"
redis = "^5.2.0"

